# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dots and Boxes Neon",
        "version": "1.0.0"
    },
    "paths": {
        "/api/ListaSalas": {
            "get": {
                "tags": [
                    "Sala"
                ],
                "summary": "Listar las Salas",
                "operationId": "getSala",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token de Firebase",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todas las salas."
                    },
                    "404": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/CrearTablero": {
            "post": {
                "tags": [
                    "Tablero"
                ],
                "summary": "Crear tablero",
                "description": "",
                "operationId": "CrearTablero",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token de Firebase",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sala_id",
                        "in": "query",
                        "description": "Id de la sala",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "x",
                        "in": "query",
                        "description": "Movimiento  X",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "y",
                        "in": "query",
                        "description": "Movimiento  Y",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ok, Juegos Disponibles, Juegos Usados"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Sala no existe"
                    }
                }
            }
        },
        "/api/ListaTablerosPorSala/{id}": {
            "get": {
                "tags": [
                    "Tablero"
                ],
                "summary": "Listar tableros por sala",
                "operationId": "getListaTablerosPorSala",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token de Firebase",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sala",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todas las tablas por sala_id."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "No hay tableros con este ID"
                    }
                }
            }
        },
        "/api/CrearJuego": {
            "post": {
                "tags": [
                    "Tablero"
                ],
                "summary": "Crear Juego",
                "description": "El Jugador 1 no puede ser Igual al jugador 2, enviar solo idTableros de otros jugadores",
                "operationId": "crearJuego",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token de Firebase",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "IdTablero",
                        "in": "query",
                        "description": "Id de la sala",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok, Juegos Disponibles, Juegos Usados"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Error j1 y j2 son el mismo jugador"
                    },
                    "422": {
                        "description": "Tablero no existe"
                    }
                }
            }
        },
        "/api/verJuego/{id}": {
            "get": {
                "tags": [
                    "Tablero"
                ],
                "summary": "Ver Juego",
                "operationId": "getVerJuego",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token de Firebase",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tablero",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar el juego mediande el ID"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Tablero no encontrado, Tablero en espera, necesita un tablero activo"
                    },
                    "422": {
                        "description": "No hay tableros con este ID"
                    }
                }
            }
        },
        "/api/Jugada": {
            "post": {
                "tags": [
                    "Tablero"
                ],
                "summary": "Jugada",
                "description": "El Jugador 1 no puede ser Igual al jugador 2, enviar solo idTableros de otros jugadores",
                "operationId": "Jugada",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token de Firebase",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "IdTablero",
                        "in": "query",
                        "description": "Id de la sala",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok, Activo, jugadorIdTurno"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Tablero No encontrado,bloqueado,Espera,Finalizado, La jugada X|Y no es valida, Este movimiento ya fue realizado"
                    },
                    "422": {
                        "description": "Tablero no existe"
                    }
                }
            }
        },
        "/api/PerfilJugador/{id}": {
            "get": {
                "tags": [
                    "Tablero"
                ],
                "summary": "Perfil Jugador",
                "operationId": "getPerfilJugador",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token de Firebase",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del jugador",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Id, Partidas, Victorias, Puntos"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Usuario no existe"
                    }
                }
            }
        },
        "/api/historialTableroId/{id}": {
            "get": {
                "tags": [
                    "Tablero"
                ],
                "summary": "Historial Tablero",
                "operationId": "getHistorialTablero",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token de Firebase",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tablero",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historial"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "No hay historial para este tablero"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}